pyproject.toml
[project]
name = "simplon-mlops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "mlflow>=3.10.0",
    "myst-parser>=5.0.0",
    "numpy>=2.4.2",
    "python-dotenv>=1.2.2",
    "sphinx-autobuild>=2025.8.25",
    "sqlalchemy>=2.0.47",
    "streamlit>=1.55.0",
    "typing>=3.10.0.0",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "furo>=2025.12.19",
    "mypy>=1.19.1",
    "myst-parser>=5.0.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.4",
    "sphinx>=9.0.4",
    "sphinx-rtd-theme>=3.1.0",
]


[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "D"]
ignore = ["D100", "D203", "D213"]

[tool.black]
line-length = 88

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=app --cov-report=term-missing"
